requirements.txt
langchain==0.3.27
openai==1.98.0
langchain-openai==0.3.28
langchain-community==0.3.27
python-dotenv==1.0.0
wikipedia==1.4.0
numexpr==2.11.0